fixed payment amount long to float

+++ src/main/java/com/ohusiev/budgettracker/persistence/model/Payment.java
@@ -6,5 +6,5 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 @Document(collection = "payments")
-public record Payment(@Id String id, Long amount, String description, LocalDateTime creationDate) {
+public record Payment(@Id String id, Double amount, String description, LocalDateTime creationDate) {
 }
